tests: Send diagnostics to stderr
If stderr and stdout are going to the same place, it doesn't matter
either way. If they are separated (as they are in the Debian
autopkgtest environment), we want the diagnostic that indicates "this
next warning is OK" to end up in the same place as the warning.

Signed-off-by: Simon McVittie <smcv@collabora.com>

Closes: #229
Approved by: cgwalters